Always resize the last chunk
This might cause errors when having a fake chunk as the very last chunk
in the code. This might happen when having some orphan asm code at the
very end.
If we don't resize it, that fake chunk will have end_addr == BADADDR, so
it will basically be unlimited. This might result in errors when binary
searching for chunks.